+++ src/index.tsx
@@ -52,8 +52,10 @@ const store = configureStore({
                     // contain ArrayBuffer or DataView
                     'data',
                     'firmwareZip',
+                    'payload',
                     // Error is not serializable
                     'error',
+                    'props.error',
                 ],
             },
         })