+++ src/toolbar/buttons/run/RunButton.tsx
@@ -16,7 +16,7 @@ const RunButton: React.VoidFunctionComponent<RunButtonProps> = ({ id }) => {
     const downloadProgress = useSelector((s) => s.hub.downloadProgress);
     const mpyAbiVersion = useSelector((s) => s.hub.mpyAbiVersion);
     const runtime = useSelector((s) => s.hub.runtime);
-    const isEditorReady = useSelector((s) => s.editor.isReady);
+    const activeFile = useSelector((s) => s.editor.activeFileUuid);
     const keyboardShortcut = 'F5';
 
     const i18n = useI18n();
@@ -35,7 +35,7 @@ const RunButton: React.VoidFunctionComponent<RunButtonProps> = ({ id }) => {
                     : i18n.translate('tooltip.action', { key: keyboardShortcut })
             }
             icon={icon}
-            enabled={isEditorReady && runtime === HubRuntimeState.Idle}
+            enabled={activeFile !== null && runtime === HubRuntimeState.Idle}
             showProgress={runtime === HubRuntimeState.Loading}
             progress={downloadProgress === null ? undefined : downloadProgress}
             onAction={() => dispatch(downloadAndRun(mpyAbiVersion))}